# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  handlers.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  handlers.LoginResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
      token_type:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: User-scoped filesystem API for containerd-backed data.
  title: Memoh Go API
  version: "1.0"
paths:
  /auth/login:
    post:
      description: Validate user credentials and issue a JWT
      parameters:
      - description: Login request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Login
      tags:
      - auth
...
schemes:
- http
swagger: "2.0"